% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{check_column_exist}
\alias{check_column_exist}
\title{Function to check the given column exists}
\usage{
check_column_exist(column_name, data)
}
\arguments{
\item{column_name}{a column name}

\item{data}{data frame}
}
\value{
0 if success -1 if failure
}
\description{
Function to check the given column exists
}
\examples{
check_column_exist("age", data.frame(
  age = rep(20, 4), sex = rep("male", 4),
  stringsAsFactors = FALSE
))
}
