% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{check_scores_3L}
\alias{check_scores_3L}
\title{Function to check the EQ-5D-3L scores}
\usage{
check_scores_3L(dimen, dimen2 = NA, dimen3 = NA, dimen4 = NA, dimen5 = NA)
}
\arguments{
\item{dimen}{a must input,response for EQ-5D-3L mobility  or the 5 digit 
response, or the vector of responses, e.g. 11111, c(1, 1, 1, 1, 1) or 1}

\item{dimen2}{response for EQ-5D-3L self care, or NA if the responses 
are given as dimensions}

\item{dimen3}{response for EQ-5D-3L usual activities,or NA if the 
responses are given as dimensions}

\item{dimen4}{response for EQ-5D-3L pain/discomfort, or NA if the 
responses are given as dimensions}

\item{dimen5}{response for EQ-5D-3L anxiety/depression, or NA if 
the responses are given as dimensions}
}
\description{
Function to check the EQ-5D-3L scores
}
\examples{
check_scores_3L(c(1, 2, 3, 3, 3))
check_scores_3L(1, 2, 3, 3, 3)
check_scores_3L(1, 2, 3, 2, 3)
check_scores_3L(12323)
}
