% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{checkColumnExist}
\alias{checkColumnExist}
\title{Function to check the given column exists}
\usage{
checkColumnExist(column.name, data)
}
\arguments{
\item{column.name}{a column name}

\item{data}{data frame}
}
\value{
0 if success -1 if failure
}
\description{
Function to check the given column exists
}
\examples{
checkColumnExist("age",data.frame(age=rep(20, 4), sex=rep("male", 4),stringsAsFactors=FALSE))
}
