% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{checkScores3L}
\alias{checkScores3L}
\title{Function to check the EQ-5D-3L scores}
\usage{
checkScores3L(dimen, dimen2 = NA, dimen3 = NA, dimen4 = NA,
  dimen5 = NA)
}
\arguments{
\item{dimen}{a must input,response for EQ-5D-3L mobility  or the 5 digit response, or the vector of responses, e.g. 11111, c(1,1,1,1,1) or 1}

\item{dimen2}{response for EQ-5D-3L self care, or NA if the responses are given as dimen}

\item{dimen3}{response for EQ-5D-3L usual activities,or NA if the responses are given as dimen}

\item{dimen4}{response for EQ-5D-3L pain/discomfort, or NA if the responses are given as dimen}

\item{dimen5}{response for EQ-5D-3L anxiety/depression, or NA if the responses are given as dimen}
}
\description{
Function to check the EQ-5D-3L scores
}
\examples{
checkScores3L(c(1, 2, 3, 3, 3))
checkScores3L(1, 2, 3, 3, 3)
checkScores3L(1, 2, 3, 2, 3)
checkScores3L(12323)
}
