% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{checkScores5L}
\alias{checkScores5L}
\title{Function to check the EQ-5D-5L scores}
\usage{
checkScores5L(dimen, dimen2 = NA, dimen3 = NA, dimen4 = NA,
  dimen5 = NA)
}
\arguments{
\item{dimen}{a must input,response for EQ-5D-3L mobility  or the 5 digit response, or the vector of responses, e.g. 11111, c(1,1,1,1,1) or 1}

\item{dimen2}{response for EQ-5D-5L  self care, or NA if the responses are given as dimen}

\item{dimen3}{response for EQ-5D-5L  usual activities,or NA if the responses are given as dimen}

\item{dimen4}{response for EQ-5D-5L  pain/discomfort, or NA if the responses are given as dimen}

\item{dimen5}{response for EQ-5D-5L  anxiety/depression, or NA if the responses are given as dimen}
}
\description{
Function to check the EQ-5D-5L scores
}
\examples{
checkScores5L(c(1, 2, 3, 5, 3))
checkScores5L(1, 2, 3, 4, 3)
checkScores5L(12323)
}
