% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{convert_number_to_digits}
\alias{convert_number_to_digits}
\title{Function to convert a number to individual digits}
\usage{
convert_number_to_digits(this_number)
}
\arguments{
\item{this_number}{a number}
}
\value{
digits
}
\description{
Function to convert a number to individual digits
}
\examples{
convert_number_to_digits(234)
}
