% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{convertNumberToIndividualDigits}
\alias{convertNumberToIndividualDigits}
\title{Function to convert a number to individual digits}
\usage{
convertNumberToIndividualDigits(this.number)
}
\arguments{
\item{this.number}{a number}
}
\value{
digits
}
\description{
Function to convert a number to individual digits
}
\examples{
convertNumberToIndividualDigits(234)
}
