% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{correctimplausibleordering}
\alias{correctimplausibleordering}
\title{Function to correct the implausible ordering in Australian valueset for EQ-5D-3L}
\usage{
correctimplausibleordering(scores)
}
\arguments{
\item{scores}{, EQ-5D-3L scores as a number}
}
\value{
the value that read from the stored dataframe
correctimplausibleordering(11121)
}
\description{
Correcting the implausible ordering
}
