% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{descriptive_stat_data_column}
\alias{descriptive_stat_data_column}
\title{Function to return descriptive statistics, sum, no of observations, 
mean, mode. median, range, standard deviation and standard error}
\usage{
descriptive_stat_data_column(colum, column_name, nrcode = NA)
}
\arguments{
\item{colum}{column}

\item{column_name}{the column name}

\item{nrcode}{non response code corresponding to the column}
}
\value{
the descriptive statistics for success , -1 for failure
}
\description{
Function to return descriptive statistics, sum, no of observations, 
mean, mode. median, range, standard deviation and standard error
}
\examples{
descriptive_stat_data_column(c(1, 2, 3, 4, NA), "scores", NA)
}
