% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{descriptiveStatDataColumn}
\alias{descriptiveStatDataColumn}
\title{Function to return descriptive statistics, sum, no of observations, mean, mode. median, range, standard deviation and standard error}
\usage{
descriptiveStatDataColumn(colum, column.name, nrcode = NA)
}
\arguments{
\item{colum}{column}

\item{column.name}{the column name}

\item{nrcode}{non response code corresponding to the column}
}
\value{
the descriptive statistics for success , -1 for failure
}
\description{
Function to return descriptive statistics, sum, no of observations, mean, mode. median, range, standard deviation and standard error
}
\examples{
descriptiveStatDataColumn(c(1,2,3,4,NA),"scores",NA)
}
