% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{.correctImplausibleOrdering}
\alias{.correctImplausibleOrdering}
\title{Function to correct the implausible ordering in Australian valueset for EQ-5D-3L}
\usage{
.correctImplausibleOrdering(scores)
}
\arguments{
\item{scores}{, EQ-5D-3L scores as a number}
}
\value{
the value that read from the stored dataframe
}
\description{
Correcting the implausible ordering
}
\examples{
.correctImplausibleOrdering(11121)
}
