% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D3L_indexvalues.df}
\alias{EQ5D3L_indexvalues.df}
\title{EQ-5D-3L index values (for each set of response of 3L) for different countries}
\format{A  243 by 28 dataframe}
\usage{
EQ5D3L_indexvalues.df
}
\description{
EQ-5D-3L index values (for each set of response of 3L) for different countries
}
\keyword{datasets}
