% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D3L_indexvalues_TTO_UK.df}
\alias{EQ5D3L_indexvalues_TTO_UK.df}
\title{EQ5D 3L index values (for each set of response of 3L) using TTO for UK}
\format{A  243 by 2 dataframe
From ref:Dolan et al. 1997 and https://www.economicsnetwork.ac.uk/health/EQ_5D_index_calculator.xls}
\usage{
EQ5D3L_indexvalues_TTO_UK.df
}
\description{
EQ5D 3L index values (for each set of response of 3L) using TTO for UK
}
\keyword{datasets}
