% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D3L_tariffs_TTO.df}
\alias{EQ5D3L_tariffs_TTO.df}
\title{EQ-5D-3L tariffs using TTO for different countries}
\format{A 21 by 20 dataframe}
\usage{
EQ5D3L_tariffs_TTO.df
}
\description{
EQ-5D-3L tariffs using TTO for different countries
}
\keyword{datasets}
