% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D3L_tariffs_VAS.df}
\alias{EQ5D3L_tariffs_VAS.df}
\title{EQ-5D-3L tariffs using VAS for different countries}
\format{A 14 by 10 dataframe}
\usage{
EQ5D3L_tariffs_VAS.df
}
\description{
EQ-5D-3L tariffs using VAS for different countries
}
\keyword{datasets}
