% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D5L_Crosswalk_ValueSets.df}
\alias{EQ5D5L_Crosswalk_ValueSets.df}
\title{EQ5D 5L cross walk value sets for UK ie values of 3L for each index value of 5L}
\format{A  3125 by 16 dataframe
From ref:https://euroqol.org/eq-5d-instruments/eq-5d-5l-about/valuation-standard-value-sets/crosswalk-index-value-calculator/}
\usage{
EQ5D5L_Crosswalk_ValueSets.df
}
\description{
EQ5D 5L cross walk value sets for UK ie values of 3L for each index value of 5L
}
\keyword{datasets}
