% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D5L_indexvalues.df}
\alias{EQ5D5L_indexvalues.df}
\title{EQ5D 5L value sets}
\format{A 5 by 7 dataframe}
\usage{
EQ5D5L_indexvalues.df
}
\description{
EQ5D 5L value sets
}
\keyword{datasets}
