% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D5L_tariffs.df}
\alias{EQ5D5L_tariffs.df}
\title{EQ-5D-5L tariffs for different countries}
\format{A 34 by 16 dataframe}
\usage{
EQ5D5L_tariffs.df
}
\description{
EQ-5D-5L tariffs for different countries
}
\keyword{datasets}
