% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5d5L_valuesets_13countries.df}
\alias{EQ5d5L_valuesets_13countries.df}
\title{EQ5D 5L value sets for all 13 countries}
\format{A 5 by 7 dataframe}
\usage{
EQ5d5L_valuesets_13countries.df
}
\description{
EQ5D 5L value sets for all 13 countries
}
\keyword{datasets}
