% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{eq5dmap5Lto3L}
\alias{eq5dmap5Lto3L}
\title{Function to map EQ-5D-5L scores to EQ-5D-3L index values as per the specific country and group by gender and age}
\usage{
eq5dmap5Lto3L(eq5dresponse.data, mobility, self.care, usual.activities,
  pain.discomfort, anxiety, country = "UK", method = "CW",
  groupby = NULL, agelimit = NULL)
}
\arguments{
\item{eq5dresponse.data}{the data containing eq5d5L responses}

\item{mobility}{column name for EQ-5D-5L mobility}

\item{self.care}{column name for response for EQ-5D-5L self care}

\item{usual.activities}{column name for response for EQ-5D-5L usual activities}

\item{pain.discomfort}{column name for response for EQ-5D-5L pain/discomfort}

\item{anxiety}{column name for response for EQ-5D-5L anxiety/depression}

\item{country}{country of interest, by default is UK, if groupby has to specify the country should be specified}

\item{method}{CW cross walk}

\item{groupby}{male or female -grouping by gender, default NULL}

\item{agelimit}{vector of ages to show upper and lower limits}
}
\value{
index value  if success, negative values for failure
}
\description{
Function to map EQ-5D-5L scores to EQ-5D-3L index values
}
\examples{
eq5dmap5Lto3L(data, "Mobility", "SelfCare","UsualActivity", "Pain", "Anxiety",UK,NULL,c(10,70))
}
