% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5Dmap_table5.df}
\alias{EQ5Dmap_table5.df}
\title{EQ-5D-5L mapping 5Lto 3L values}
\format{
A 31250 by 21 dataframe
}
\source{
Hernandez Alava et al (2017) <doi.org/10.1016/j.jval.2016.11.006>
}
\usage{
EQ5Dmap_table5.df
}
\description{
EQ-5D-5L mapping 5Lto 3L values
}
\keyword{datasets}
