% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{get_colno_existing_colnames}
\alias{get_colno_existing_colnames}
\title{Function to return the column number for a given column name 
(from list of possible column names that may
have used) in a data frame}
\usage{
get_colno_existing_colnames(column_names, data)
}
\arguments{
\item{column_names}{column names in a data frame}

\item{data}{a data frame}
}
\value{
the column number
}
\description{
Function to return the column number for a given column name 
(from list of possible column names that may
have used) in a data frame
}
\examples{
get_colno_existing_colnames(c("age"), data.frame(age = rep(20, 4), 
gender = rep("male", 4)))
}
