% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{get_column_no_colnames}
\alias{get_column_no_colnames}
\title{Function to return the column number for column name}
\usage{
get_column_no_colnames(data, column_name)
}
\arguments{
\item{data}{a data frame}

\item{column_name}{column names of the data frame}
}
\value{
column number, if success -1, if failure
}
\description{
Function to return the column number for column name
}
\examples{
get_column_no_colnames(data.frame(age = rep(20, 4), 
sex = rep("male", 4)), "sex")
}
