% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{get_frequency_table}
\alias{get_frequency_table}
\title{Function to return frequency table}
\usage{
get_frequency_table(v)
}
\arguments{
\item{v}{a vector}
}
\value{
frequency table
}
\description{
Function to return frequency table
}
\examples{
get_frequency_table(c(1, 1, 1, 12, 2))
}
