% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{get_mode_for_vec}
\alias{get_mode_for_vec}
\title{Function to return mode}
\usage{
get_mode_for_vec(v)
}
\arguments{
\item{v}{a vector}
}
\value{
mode if success -1 for failure
}
\description{
Function to return mode
}
\examples{
get_mode_for_vec(c(1, 1, 2, 3))
}
