% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{getColNumExistingColNames}
\alias{getColNumExistingColNames}
\title{Function to return the column number for a given column name (from list of possible column names that may
have used) in a data frame}
\usage{
getColNumExistingColNames(column.names, data)
}
\arguments{
\item{column.names}{column names in a data frame}

\item{data}{a data frame}
}
\value{
the column number
}
\description{
Function to return the column number for a given column name (from list of possible column names that may
have used) in a data frame
}
\examples{
getColNumExistingColNames(c("age"), data.frame(age = rep(20, 4), gender = rep("male", 4)))
}
