% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{getColumnNoColNames}
\alias{getColumnNoColNames}
\title{Function to return the column number for column name}
\usage{
getColumnNoColNames(data, column.name)
}
\arguments{
\item{data}{a data frame}

\item{column.name}{column names of the data frame}
}
\value{
column number, if success -1, if failure
}
\description{
Function to return the column number for column name
}
\examples{
getColumnNoColNames(data.frame(age=rep(20, 4), sex=rep("male", 4)),"sex")
}
