% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{getModeForVec}
\alias{getModeForVec}
\title{Function to return mode}
\usage{
getModeForVec(v)
}
\arguments{
\item{v}{a vector}
}
\value{
mode if success -1 for failure
}
\description{
Function to return mode
}
\examples{
getModeForVec(c(1,1,2,3))
}
