% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{map_5Lto3L_Ind_NICE2022}
\alias{map_5Lto3L_Ind_NICE2022}
\title{Function to map EQ-5D-5L descriptive system to 3L index value
using  Hernandez et al (2017) method and DSU's functional approach for
NICE guidance 2022
countries are UK(England), Japan, Korea, Netherlands, China, Spain and Germany}
\usage{
map_5Lto3L_Ind_NICE2022(
  country,
  gender,
  dimen = NA,
  age = NA,
  agegroup = NA,
  dimen1 = NA,
  dimen2 = NA,
  dimen3 = NA,
  dimen4 = NA,
  dimen5 = NA
)
}
\arguments{
\item{country}{specific country from the above country list}

\item{gender}{gender}

\item{dimen}{response for EQ-5D-5L mobility  or the 5 digit response, or 
the vector of responses, e.g. 11111, c(1,1,1,1,1) or 1}

\item{age}{age if given as exact age}

\item{agegroup}{if age is not known age groupsshould be given
they are 16-35, 35-45, 45-55, 55-65, 65-100. 1-5}

\item{dimen1}{response for EQ-5D-5L mobility, or NA if the responses are 
given as dimen}

\item{dimen2}{response for EQ-5D-5L self care, or NA if the responses are 
given as dimen}

\item{dimen3}{response for EQ-5D-5L usual activities,or NA if the responses
are given as dimen}

\item{dimen4}{response for EQ-5D-5L pain/discomfort, or NA if the responses
are given as dimen}

\item{dimen5}{response for EQ-5D-5L anxiety/depression, or NA if the 
responses are given as dimen}
}
\value{
index value of EQ-5D-3L, -1 if any error
}
\description{
Function to map EQ-5D-5L descriptive system to 3L index value
(ref:Hernandez, M. and Pudney, S. (2017) and code inspired from 
https://www.sheffield.ac.uk/nice-dsu/methods-development/mapping-eq-5d-5l-3l)
}
\examples{
map_5Lto3L_Ind_NICE2022("England", "female", 11121, 30)
map_5Lto3L_Ind_NICE2022("England", "female", NA, 30, NA, 1,2,3,4,5)
}
