% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{replace_space_underscore}
\alias{replace_space_underscore}
\title{Function to add an underscore for texts with spaces in between}
\usage{
replace_space_underscore(this_string)
}
\arguments{
\item{this_string}{a string}
}
\value{
string where the spaces replaced by "_"
}
\description{
Function to add an underscore for texts with spaces in between
}
\examples{
replace_space_underscore("Sri Lanka")
}
