% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{replaceSpaceUnderscore}
\alias{replaceSpaceUnderscore}
\title{Function to add an underscore for texts with spaces in between}
\usage{
replaceSpaceUnderscore(this.string)
}
\arguments{
\item{this.string}{a string}
}
\value{
string where the spaces replaced by "_"
}
\description{
Function to add an underscore for texts with spaces in between
}
\examples{
replaceSpaceUnderscore("Sri Lanka")
}
