% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{subsetGenderAgeToGroup}
\alias{subsetGenderAgeToGroup}
\title{Function to check the gender column and age column subset based on the values in it
have used) in a data frame}
\usage{
subsetGenderAgeToGroup(data, gender, agelimit)
}
\arguments{
\item{data}{a data frame}

\item{gender}{groupby gender either male or female expected}

\item{agelimit}{list of ages e.g. c(10,20)}
}
\value{
the column number
}
\description{
Function to check the gender column and age column subset based on the values in it
have used) in a data frame
}
\examples{
subsetGenderAgeToGroup(data.frame(age = rep(20, 4), gender = rep("male", 4)), "male", c(10, 70))
}
