% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{test_data_num_norange}
\alias{test_data_num_norange}
\title{Function to check format of a numeric column when the values are not bounded}
\usage{
test_data_num_norange(vec, nrcode = NA)
}
\arguments{
\item{vec}{a column vector}

\item{nrcode}{non response code corresponding to the column}
}
\value{
0, if success -1, if failure
}
\description{
Function to check format of a numeric column when the values are not bounded
}
\examples{
test_data_num_norange(c(1, 2, 3, 4, -99), -99)
}
