% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{test_file_exist_read}
\alias{test_file_exist_read}
\title{Function to throw error on invalid directory or file or if the file is 
not readable}
\usage{
test_file_exist_read(filename)
}
\arguments{
\item{filename}{name of a file or directory}
}
\value{
0 if success, non zero negative values if failure
}
\description{
Function to throw error on invalid directory or file or if the file is 
not readable
}
\examples{
test_file_exist_read(system.file("extdata", "blank.txt", 
package = "valueEQ5D"))
}
