% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{testDataNumNorange}
\alias{testDataNumNorange}
\title{Function to check the format of a numeric column when the values are not bounded}
\usage{
testDataNumNorange(vec, nrcode = NA)
}
\arguments{
\item{vec}{a column vector}

\item{nrcode}{non response code corresponding to the column}
}
\value{
0, if success -1, if failure
}
\description{
Function to check the format of a numeric column when the values are not bounded
}
\examples{
testDataNumNorange(c(1,2,3,4,-99),-99)
}
