% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{testFileExistRead}
\alias{testFileExistRead}
\title{Function to throw error on invalid directory or file or if the file is not readable}
\usage{
testFileExistRead(filename)
}
\arguments{
\item{filename}{name of a file or directory}
}
\value{
0 if success, non zero negative values if failure
}
\description{
Function to throw error on invalid directory or file or if the file is not readable
}
\examples{
testFileExistRead("smk543/Desktop/MKrishnan/Project I-WOTCH/Datasets/testing.csv")
}
