% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{valueEQ5D3L}
\alias{valueEQ5D3L}
\title{Function to value EQ-5D-3L columns to index values for any country and group by gender and age}
\usage{
valueEQ5D3L(eq5dresponse.data, mobility, self.care, usual.activities,
  pain.discomfort, anxiety, country, method, groupby, agelimit)
}
\arguments{
\item{eq5dresponse.data}{the data containing eq5d responses}

\item{mobility}{column name for EQ-5D-3L mobility}

\item{self.care}{column name for response for EQ-5D-3L self care}

\item{usual.activities}{column name for response for EQ-5D-3L usual activities}

\item{pain.discomfort}{column name for response for EQ-5D-3L pain/discomfort}

\item{anxiety}{column name for response for EQ-5D-3L anxiety/depression}

\item{country}{country of interest, by default is UK, if groupby has to specifiy the country should be specified}

\item{method}{Either "TTO" or "VAS"}

\item{groupby}{male or female -grouping by gender, default NULL}

\item{agelimit}{vector of ages to show upper and lower limits}
}
\value{
the descriptive statistics of index values, frequence table and the modified data where the last column will be the index values
}
\description{
Main function to value EQ-5D-5L descriptive system to 5L index values.
}
\examples{
valueEQ5D3L(data, "mo", "sc","ua", "pd", "ad","UK","TTO",NULL,c(10,70))
}
