% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{valueEQ5D5L}
\alias{valueEQ5D5L}
\title{Function to value EQ-5D-5L scores for any country and group by gende and age}
\usage{
valueEQ5D5L(eq5dresponse.data, mobility, self.care, usual.activities,
  pain.discomfort, anxiety, country = "England", groupby = NULL,
  agelimit = NULL)
}
\arguments{
\item{eq5dresponse.data}{the data containing eq5d responses}

\item{mobility}{column name for EQ-5D-5L mobility}

\item{self.care}{column name for response for EQ-5D-5L self care}

\item{usual.activities}{column name for response for EQ-5D-5L usual activities}

\item{pain.discomfort}{column name for response for EQ-5D-5L pain/discomfort}

\item{anxiety}{column name for response for EQ-5D-5L anxiety/depression}

\item{country}{country of interest, by default is England}

\item{groupby}{male or female -grouping by gender, default NULL}

\item{agelimit}{vector of ages to show upper and lower limits}
}
\value{
index value  if success, -1 if failure
}
\description{
Function to value EQ-5D-5L descriptive system to index value.
}
\examples{
valueEQ5D5L(data, "Mobility", "SelfCare","UsualActivity", "Pain", "Anxiety",UK,NULL,c(10,70))
}
