% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{checkScores3L}
\alias{checkScores3L}
\title{Function to check the EQ-5D-3L scores}
\usage{
checkScores3L(this.response, this.response2 = NA, this.response3 = NA,
  this.response4 = NA, this.response5 = NA)
}
\arguments{
\item{this.response}{a must input,response for EQ-5D-3L mobility  or the 5 digit response, or the vector of responses, e.g. 11111, c(1,1,1,1,1) or 1}

\item{this.response2}{response for EQ-5D-3L self care, or NA if the responses are given as this.response}

\item{this.response3}{response for EQ-5D-3L usual activities,or NA if the responses are given as this.response}

\item{this.response4}{response for EQ-5D-3L pain/discomfort, or NA if the responses are given as this.response}

\item{this.response5}{response for EQ-5D-3L anxiety/depression, or NA if the responses are given as this.response}
}
\description{
Function to check the EQ-5D-3L scores
}
\examples{
checkScores3L(c(1,2,3,3,3))
checkScores3L(1,2,3,3,3)
checkScores3L(1,2,3,2,3)
checkScores3L(12323)
}
