% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D3L_tariffs_TTO.df}
\alias{EQ5D3L_tariffs_TTO.df}
\title{EQ-5D-3L tariffs using TTO for different countries}
\format{A 38 by 25 dataframe}
\source{
\strong{Argentina}: Table 3 column 5 page 560 in Augustovski et al (2009) <doi:10.1111/j.1524-4733.2008.00468.x>

\strong{Australia}: Table 4 column 6 page 933 in Viney et al (2011) <doi:10.1016/j.jval.2011.04.009>

\strong{Brazil}: Table 2 column 8 page 21 in Santos et al (2016) <doi:10.1177/0272989X15613521>

\strong{Canada}: Table 4 column 2 page 8 in Bansback et al (2012) <https://doi.org/10.1371/journal.pone.0031115>

\strong{Chile}: Table 2 column 5 page 1137 in Zarate et al (2011) <doi:10.1016/j.jval.2011.09.002

\strong{China}: Table 4 column 4 page 603 in Liu et al (2014) <doi:10.1016/j.jval.2014.05.007>

\strong{Denmark}: Table 4 column 2 page 463 in Wittrup-Jensen et al (2009) <doi:10.1177/1403494809105287>

\strong{France}: Equation page 61 in Chevalier et al (2013) <doi:10.1007/s10198-011-0351-x>

\strong{Germany}: Table 4 column 2 page 129 in Greiner et al (2005) <doi:10.1007/s10198-004-0264-z>

\strong{Italy}: Table 4 column 5 page 820 in Scalone et al (2013) <http://dx.doi.org/10.1016/j.jval.2013.04.008>

\strong{Japan}: Table 4 column 1 page 41 in Tsuchiya et al (2002) <https://doi.org/10.1002/hec.673>

\strong{South Korea}: Table 3 column 4 page 1191 in Lee et al <doi:10.1111/j.1524-4733.2009.00579.x>

\strong{Netherlands}: Table 5 column 3 page 1128 in Lamers et al <doi:10.1002/hec.1124>

\strong{Poland}: Table 5 column 2 page 293 in Golicki et al <https://doi.org/10.1111/j.1524-4733.2009.00596.x>

\strong{Portugal}: Table 4 column 6 page 418 in Ferreira et al <doi:10.1007/s11136-013-0448-z>

\strong{Singapore}: Equation page  504 in Nan Luoß et al <doi:10.1007/s40273-014-0142-1>

\strong{Spain}: Table 3 column 4 page 13 in Badia et al (2001) <doi:10.1177/0272989X0102100102>

\strong{ Sri Lanka}: Table 2 column 8 page 1791 in  Kularatna et al (2015) <doi:10.1007/s11136-014-0906-2>

\strong{Sweden}: Table 2 column 8 page 436 in Burström et al (2014) <doi:10.1007/s11136-013-0496-4>

\strong{Taiwan}: Table 2 column 4 page 702 in Lee et al (2013) <http://dx.doi.org/10.1016/j.jfma.2012.12.015> #'

\strong{Thailand}: Table 1 column 2 page 1144 (parameters like MO3 are calculated) Tongsiri et al (2011) <doi:10.1016/j.jval.2011.06.005>

\strong{Trinidad and Tobago}: Table 4 page 65 in Bailey et al (2016) <http://dx.doi.org/10.1016/j.vhri.2016.07.010>

\strong{UK}: Table 1 column 2 page 1103 in Dolan et al (1997) <http://dx.doi.org/10.1097/00005650-199711000-00002>

\strong{USA}: Table 5 column 2 page 214 in Shaw et al (2005) <doi:10.1097/00005650-200503000-00003>

\strong{Zimbabwe}: Table 5 column 3 page 7 in Jelsma et al (2003) <https://doi.org/10.1186/1478-7954-1-11>
}
\usage{
EQ5D3L_tariffs_TTO.df
}
\description{
EQ-5D-3L tariffs using TTO for different countries
}
\keyword{datasets}
