% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D3L_tariffs_TTO.df}
\alias{EQ5D3L_tariffs_TTO.df}
\title{EQ5D 5L value sets for all 13 countries}
\format{A 5 by 7 dataframe}
\usage{
EQ5D3L_tariffs_TTO.df
}
\description{
EQ5D 5L value sets for all 13 countries
}
\keyword{datasets}
