% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D3L_tariffs_VAS.df}
\alias{EQ5D3L_tariffs_VAS.df}
\title{EQ-5D-3L tariffs using VAS for different countries}
\format{
A 34 by 12 dataframe
}
\source{
\strong{Argentina}: Table 3 column 2 page 560 in Augustovski et al (2009) <doi:10.1111/j.1524-4733.2008.00468.x>

\strong{Belgium}: Equation 2 page 208 in Cleemput et al (2010) <doi:10.1007/s10198-009-0167-0>

\strong{Denmark}: Table 2.3 page 14 in Szende et al (2014) <doi:10.1007/978-94-007-7596-1>

\strong{Europe}: Table 2.3 page 14 in Szende et al (2014) <doi:10.1007/978-94-007-7596-1>

\strong{Finland}: Table 2.3 page 14 in Szende et al (2014) <doi:10.1007/978-94-007-7596-1>

\strong{Germany}: Table 2.3 page 14 in Szende et al (2014) <doi:10.1007/978-94-007-7596-1>

\strong{Malaysia}: Table 4 column 6 page S88 in Yusof et al (2019) <doi:10.1016/j.jval.2011.11.024>

\strong{New Zealand}: Equation 2 page 541 in Devlin et al <doi:10.1002/hec.741>

\strong{Slovenia}: Table 2.3 page 14 in Szende et al (2014) <doi:10.1007/978-94-007-7596-1>

\strong{Spain}: Table 2.3 page 14 in Szende et al (2014) <doi:10.1007/978-94-007-7596-1> (this is not shown in euroqol website)
Could not get Sweden VAS values

\strong{UK}: Table 2.3 page 14 in Szende et al (2014) <doi:10.1007/978-94-007-7596-1>
}
\usage{
EQ5D3L_tariffs_VAS.df
}
\description{
EQ-5D-3L tariffs using VAS for different countries
}
\keyword{datasets}
