% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D5L_crosswalk_indexvalues.df}
\alias{EQ5D5L_crosswalk_indexvalues.df}
\title{EQ-5D-5L crosswalk value sets for 10 countries}
\format{A 3125 by 11 dataframe}
\source{
https://euroqol.org/eq-5d-instruments/eq-5d-5l-about/valuation-standard-value-sets/crosswalk-index-value-calculator/ (accessed on Aug 03,2019)

Van Hout et al (2012) <doi: 10.1016/j.jval.2012.02.008>.
}
\usage{
EQ5D5L_crosswalk_indexvalues.df
}
\description{
EQ-5D-5L crosswalk value sets for 10 countries
}
\note{
: For testing purpose -not required by users
}
\keyword{datasets}
