% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D5L_crosswalk_indexvalues.df}
\alias{EQ5D5L_crosswalk_indexvalues.df}
\title{EQ-5D-5L crosswalk value sets for 10 countries}
\format{A 3125 by 11 dataframe}
\usage{
EQ5D5L_crosswalk_indexvalues.df
}
\description{
EQ-5D-5L crosswalk value sets for 10 countries
}
\keyword{datasets}
