% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D5L_indexvalues.df}
\alias{EQ5D5L_indexvalues.df}
\title{EQ-5D-5L index values
 @note: For testing purpose -not required by users}
\format{A 3125 by 18 dataframe}
\source{
\strong{Canada}: Selected example Table A3 in Xie et al (2016) <doi:10.1097/MLR.0000000000000447>

\strong{China}: Supplementary Material 1 in Luo et al (2017) <doi:10.1016/j.jval.2016.11.016>

\strong{England}: Selected example Table 3 page 18 and supporting information in Devlin et al (2018) <doi:10.1002/hec.3564>

\strong{Germany}: Supplementary Material 1 in Ludwig et al (2018) <doi:10.1007/s40273-018-0615-8>

\strong{Hong Kong}: Selected examples Table 3 page 244 in Wong et al (2018) <doi:10.1007/s40271-017-0278-0>

\strong{Indonesia}: Selected examples page 1162 in Purba et al (2017) <doi:10.1007/s40273-017-0538-9>

\strong{Ireland}: Selected example Table 2 page 1348 in Hobbins et al (2016) <doi:10.1007/s40273-018-0690-x>

\strong{Japan}: Shiroiwa, et al (2016) <doi:10.1016/j.jval.2016.03.1834>

\strong{Korea}: Selected example page 1848 in Kim et al (2016) <doi:10.1007/s11136-015-1205-2>

\strong{Malaysia}: Shafie  et al (2019) <doi:10.1007/s40273-018-0758-7>

\strong{Netherlands}: Versteegh et al (2016) <doi:10.1016/j.jval.2016.01.003>

\strong{Poland}: Supplementary Material 3 in Golicki et al <doi:10.1007/s40273-019-00811-7>

\strong{Portugal}: Ferreira1 et al (2014) <doi:10.1007/s11136-019-02226-5>

\strong{Spain}: Selected examples in Table 1 in Ramos-Goñiet et al (2018) <https://doi.org/10.1016/j.jval.2017.10.023>

\strong{Taiwan}: Table 3 page 10 in Lin et al (2018)  <https://doi.org/10.1371/journal.pone.0209344>

\strong{Thailand}: Pattanaphesaj et al (2018) <doi:10.1080/14737167.2018>

\strong{Uruguay}: Augustovski et al (2016) <doi:10.1007/s11136-015-1086-4>
}
\usage{
EQ5D5L_indexvalues.df
}
\description{
EQ-5D-5L index values
 @note: For testing purpose -not required by users
}
\keyword{datasets}
