% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D5L_indexvalues.df}
\alias{EQ5D5L_indexvalues.df}
\title{EQ-5D-5L index values}
\format{A 3125 by 16 dataframe}
\usage{
EQ5D5L_indexvalues.df
}
\description{
EQ-5D-5L index values
}
\keyword{datasets}
