% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D5L_tariffs.df}
\alias{EQ5D5L_tariffs.df}
\title{EQ-5D-5L tariffs for different countries}
\format{A 34 by 17 dataframe}
\source{
\strong{Canada}: Table 2 column 5 page 103 in Xie et al (2016) <doi:10.1097/MLR.0000000000000447>

\strong{China}: Table 4 column 4 page 667 in Luo et al (2017) <doi:10.1016/j.jval.2016.11.016>

\strong{England}: Table 2 column 2 page 17 in Devlin et al (2018) <doi:10.1002/hec.3564>

\strong{Germany}: Table column 9 page 670 in Ludwig et al (2018) <doi:10.1007/s40273-018-0615-8>

\strong{Hong Kong}: Table 3 column 8 page 244 in Wong et al (2018) <doi:10.1007/s40271-017-0278-0>

\strong{Indonesia}: Table 3 column 8 page 1162 in Purba et al (2017) <doi:10.1007/s40273-017-0538-9>

\strong{Ireland}: Table 2 column 2 page 1348 in Hobbins et al (2016) <doi:10.1007/s40273-018-0690-x>

\strong{Japan}: Table 2 column 7 page 651 in Shiroiwa, et al (2016) <doi:10.1016/j.jval.2016.03.1834>

\strong{Korea}: Table 5 column 6 page 1851 in Kim et al (2016) <doi:10.1007/s11136-015-1205-2>

\strong{Malaysia}: Table 2 column 9 page 720 in Shafie  et al (2019) <doi:10.1007/s40273-018-0758-7>

\strong{Netherlands}: Table 4 column 8 page 350 in Versteegh et al (2016) <doi:10.1016/j.jval.2016.01.003>

\strong{Poland}: Table 2 column 7 in Golicki et al <doi:10.1007/s40273-019-00811-7>

\strong{Portugal}: Table 3 column 4 in Ferreira1 et al (2014) <doi:10.1007/s11136-019-02226-5>

\strong{Spain}: Table 1 column 9 page 5 in Ramos-Goñiet et al (2018) <https://doi.org/10.1016/j.jval.2017.10.023>

\strong{Taiwan}: Table 2 column 4 page 9 in Lin et al (2018)  <https://doi.org/10.1371/journal.pone.0209344>

\strong{Thailand}: Table 3 column 6 page 4 in Pattanaphesaj et al (2018) <doi:10.1080/14737167.2018>

\strong{Uruguay}: Table 2.3 column 5 page 29 in Augustovski et al (2016) <doi:10.1007/s11136-015-1086-4>
}
\usage{
EQ5D5L_tariffs.df
}
\description{
EQ-5D-5L tariffs for different countries
}
\keyword{datasets}
