% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_checks.R
\name{getFrequencyTable}
\alias{getFrequencyTable}
\title{Function to return frequency table}
\usage{
getFrequencyTable(v)
}
\arguments{
\item{v}{a vector}
}
\value{
frequency table
}
\description{
Function to return frequency table
}
\examples{
getFrequencyTable(c(1,1,1,12,2))
}
