% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{valueEQ5D3LIndscores}
\alias{valueEQ5D3LIndscores}
\title{Function to value EQ-5D-3L scores for various countries}
\usage{
valueEQ5D3LIndscores(country, method, this.response, this.response2 = NA,
  this.response3 = NA, this.response4 = NA, this.response5 = NA)
}
\arguments{
\item{country}{a country name from the list Belgium,Brazil,Canada,Chile,Denmark,Europe,Finland,France,Germany,Italy,Japan,Korea,Netherlands,NewZealand,Poland,Portugal,Slovenia,Spain,Taiwan,Thailand,UK,USA,and Zimbabwe}

\item{method}{method name either TTO or VAS}

\item{this.response}{a must input,response for EQ-5D-5L mobility  or the 5 digit response, or the vector of responses, e.g. 11111, c(1,1,1,1,1) or 1}

\item{this.response2}{response for EQ-5D-3L self care, or NA if the responses are given as this.response}

\item{this.response3}{response for EQ-5D-3L usual activities,or NA if the responses are given as this.response}

\item{this.response4}{response for EQ-5D-3L pain/discomfort, or NA if the responses are given as this.response}

\item{this.response5}{response for EQ-5D-3L anxiety/depression, or NA if the responses are given as this.response}
}
\value{
index value based if success, -1 if failure
}
\description{
Function to value EQ-5D-3L scores for various countries
}
\examples{
valueEQ5D3LIndscores("UK","TTO",23131)
valueEQ5D3LIndscores("Spain","TTO",2,3,1,3,1)
valueEQ5D3LIndscores("Denmark","VAS",c(1,2,3,1,3))
}
\references{
Argentina TTO and VAS: Table 3 column 2 and 5 for VAS and TTO page 560 in Augustovski et al (2009) <doi:10.1111/j.1524-4733.2008.00468.x> 
VAS value for state 3333 was reported as -0.022,but did not get that and answer got is -0.034

Australia: Table 4 column 6 page 933 in Viney et al (2011) <doi:10.1016/j.jval.2011.04.009>
There were some implausible orderings and hard coded those only for Australian value sets

Belgium: Equation 2 page 208 in Cleemput et al (2010) <doi:10.1007/s10198-009-0167-0>

Brazil: Table 2 column 8 page 21 in Santos et al (2016) <doi:10.1177/0272989X15613521>

Canada: Table 4 column 2 page 8 in Bansback et al (2012) <https://doi.org/10.1371/journal.pone.0031115>

Chile: Table 2 column 5 page 1137 in Zarate et al (2011) <doi:10.1016/j.jval.2011.09.002.

China: Table 4 column 4 page 603 in Liu et al (2014) <doi:10.1016/j.jval.2014.05.007>

Denmark TTO: Table 4 column 2 page 463 in Wittrup-Jensen et al (2009) <doi:10.1177/1403494809105287>

Denmark VAS: Table 2.3 page 14 in Szende et al (2014) <doi:10.1007/978-94-007-7596-1>

Europe: Table 2.3 page 14 in Szende et al (2014) <doi:10.1007/978-94-007-7596-1>

Finland: Table 2.3 page 14 in Szende et al (2014) <doi:10.1007/978-94-007-7596-1>

France: Equation page 61 in Chevalier et al (2013) <doi:10.1007/s10198-011-0351-x>

Germany TTO: Table 4 column 2 page 129 in Greiner et al (2005) <doi:10.1007/s10198-004-0264-z.

Germany VAS: Table 2.3 page 14 in Szende et al (2014) <doi:10.1007/978-94-007-7596-1>

Italy: Table 4 column 5 page 820 in Scalone et al (2013) <http://dx.doi.org/10.1016/j.jval.2013.04.008>

Japan: Table 4 column 1 page 41 in Tsuchiya et al (2002) <https://doi.org/10.1002/hec.673>

Malaysia: Table 4 column 6 page S88 in Yusof et al (2019) <doi:10.1016/j.jval.2011.11.024>

South Korea: Table 3 column 4 page 1191 in Lee et al <doi:10.1111/j.1524-4733.2009.00579.x>

Netherlands: Table 5 column 3 page 1128 in Lamers et al <doi:10.1002/hec.1124>

New Zealand: Equation 2 page 541 in Devlin et al <doi:10.1002/hec.741>

Poland: Table 5 column 2 page 293 in Golicki et al <https://doi.org/10.1111/j.1524-4733.2009.00596.x>

Portugal: Table 4 column 6 page 418 in Ferreira et al <doi:10.1007/s11136-013-0448-z>

Singapore: Equation page  504 in Nan Luoß et al <doi:10.1007/s40273-014-0142-1>

Slovenia: Table 2.3 page 14 in Szende et al (2014) <doi:10.1007/978-94-007-7596-1>

Spain TTO: Table 3 column 4 page 13 in Badia et al (2001) <doi:10.1177/0272989X0102100102>

Spain VAS: Table 2.3 page 14 in Szende et al (2014) <doi:10.1007/978-94-007-7596-1>

Sri Lanka: Table 2 column 8 page 1791 in  Kularatna et al (2015) <doi:10.1007/s11136-014-0906-2>

Sweden: Table 2 column 8 page 436 in Burström et al (2014) <doi:10.1007/s11136-013-0496-4>

Taiwan: Table 2 column 4 page 702 in Lee et al (2013) <http://dx.doi.org/10.1016/j.jfma.2012.12.015>

Thailand: Table 1 column 2 page 1144 (parameters like MO3 are calculated) Tongsiri et al (2011) <doi:10.1016/j.jval.2011.06.005>

Trinidad and Tobago: Table 4 page 65 in Bailey et al (2016) <http://dx.doi.org/10.1016/j.vhri.2016.07.010>

UK TTO: Table 1 column 2 page 1103 in Dolan et al (1997) <http://dx.doi.org/10.1097/00005650-199711000-00002>

UK VAS: Table 2.3 page 14 in Szende et al (2014) <doi:10.1007/978-94-007-7596-1>

USA: Table 5 column 2 page 214 in Shaw et al (2005) <doi:10.1097/00005650-200503000-00003>

Zimbabwe: Table 5 column 3 page 7 in Jelsma et al (2003) <https://doi.org/10.1186/1478-7954-1-11>
}
